b=22419 Restore lost mkfs_lustre.c hunks.
i=ricardo.m.correia

It appears there are a few hunks where which conflicted and were
resolved incorrectly in merge c304dfc6 of master in to b_hd_kdmu.
In particular the one which caused me trouble was a hunk which was
dropped from mkfs_lustre.c for validating correct mount options.
This commit correctly replaces the code which was lost. But the
following files also had conflicts during this update and I have
not yet looked at them closely to see that everything is OK.

Conflicts:
        lustre/ChangeLog
        lustre/autoconf/lustre-version.ac
        lustre/lclient/.gitignore
        lustre/tests/sanity-quota.sh
        lustre/tests/test-framework.sh
        lustre/utils/mkfs_lustre.c