Disable sanity test 57a (verify MDS filesystem created with large inodes) when using ZFS backend.
This test won't work with ZFS since it's ldiskfs-specific at the moment, but
even if we could find out the dnode size it wouldn't pass until we have large
dnode support. For now I'm disabling it because it's a known failure with a
known root cause and the fix is well understood.