b=22586 restore removed configure
i=ricardo.m.correia

In landing bug 19422 a configure pass was removed as it was determined at
the time to be superfluous.  It turns out that for some situations at least
it's not.  So restore it.
Refactoring all of this is in fact being handled in another bug, so this
whole area of code will be addressed, eventually, when that lands.