b=22246 Fix statfs block size and re-enable grant reservation.
The DMU's statfs implementation was returning the "optimal transfer block size"
(f_bsize) as 128K. However, there are places where f_bsize is interpreted as the
unit for f_{blocks,bfree,bavail}, instead of f_frsize, which the DMU sets as 512
bytes. The result is that statfs() was returning a much larger filesystem size
than what it was supposed to.

Also, this patch fixes and re-enables the space reservation made to avoid
running into ENOSPC due to grants not accounting for metadata overhead in ZFS.

The reservation was buggy because it only subtracted blocks from f_bsize,
leading to some cases where f_bfree and/or f_bavail could be larger than
f_bsize and cause 'df' to report negative values.