b=23267 Fix compilation errors in udmu.c due to missing include in lustre_idl.h.
It seems that lustre_idl.h uses definitions from libcfs.h (such as LASSERT,
CLASSERT, LPUX64, ...) but it doesn't include it.

Since I'm not sure whether it's OK to include libcfs.h in lustre_idl.h,
I'm including it in udmu.c as a temporary fix to get things working.