b=23709 Check module directory for symbols
i=ricardo.correia

Previously the SPL_OBJ and ZFS_OBJ directories were being
incorrectly set to the spl/module or zfs/module subdirectory.
The right thing to do here is to set it to the top level build
directory and check for the needed symbol information in
the correct subdirectories.  Some extra checks are needed
because the location of the module symbols differs slightly
depending on if we are building against installed devel
headers or against in-tree spl/zfs sources and objects.