b=22962 Don't register the space accounting callback in Lustre.
The space accounting callback is needed for any ZPL dataset, even when the
Lustre modules are not even loaded.

During testing, I've ran into an oops where the callback was registered,
afterwards the Lustre modules were unloaded, and then a ZFS filesystem was
created. The latter operation caused a crash because the Lustre callback
was still registered and it was being called even though the function wasn't in
memory anymore.

This callback must be available even when Lustre isn't running, so the proper
fix is to just use the ZPL callback and to register it in zfs_init() in Linux as
it's done in Solaris. This means that Lustre doesn't need to register the
callback anymore, so we can just remove that code.