Remove mutex_lock()/mutex_unlock() in udmu_wait_synced().
The correct API to use should have been the Solaris mutex_enter()/mutex_exit()
rather than the Linux one. The compiler didn't flag errors in some
configurations, but it caused a compilation failure in others.

Besides, the mutex is not required because it doesn't protect tx_open_txg, so
we'll remove those calls. The DMU itself also accesses it without any locking.