b=24037 Include backing-dev.h header
With ldiskfs disabled we fail to include the backing-dev.h header
which is required for the lustre_sb_info structure.  Presumably,
this was not as issue previously because it was included through
one of the ldiskfs headers.  The backing-dev.h header exists all
the way back to 2.6.0 kernels so I have included it unconditionally
in the linux/linux-fs.h header.  This header is included as part
of libcfs.h which is in turn included by lustre_disk.h.

  In file included from
  /home/behlendo/src/git/lustre-kdmu/lustre/ofd/ofd_internal.h:12, from
  /home/behlendo/src/git/lustre-kdmu/lustre/ofd/ofd_io.c:44:
  /home/behlendo/src/git/lustre-kdmu/lustre/include/lustre_disk.h:444:
  error: field ‘bdi’ has incomplete type

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>